Added functionality to build a model of Information Module Definitions.
The IMD building process handles data found in chapter C of the DICOM standard part 03. Internal helper functions were added to determine the sequence item depth and attribute name. Sequence item depth is determined by the number of larger than characters at the front of the attribute name. There are two types of IMD entries [include and data]. Each type is selected based on the number of columns and by the value of the first column. If the first column includes a link, it's considered an IncludeEntry. Otherwise, a DataEntry.

The properties in the DicomStandard class are now writable but the Lists are not mutable since the data is only modified by setting the whole list of CIOD or IMD.

The DICOM attribute type is now a separate Enum.

Unit tests and the corresponding test resources have been extended to (partially) cover the new functionality.